% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocASN.R
\name{LocASN}
\alias{LocASN}
\title{Single-cell RNA sequencing normalization using a local average technique}
\usage{
LocASN(countmatrix, conditions = NULL, filter = FALSE, 
       gene_num_gezero = 3, cell_num_gezero = 10, 
       numGeneforEst = 2000, divideforFast = FALSE, numDivide = 2)
}
\arguments{
\item{countmatrix}{Input. Unnormalized count (sparse) matrix (genes by cells).}

\item{conditions}{Input (Optional). Indicate which cells are sampled from the same conditions.
The default value, NULL, denotes all the cells are sampled from the same condition.}

\item{filter}{Input (Optional). A logic value to indicate if need data filtering. If yes, please see the details of gene_num_gezero and cell_num_gezero for input. The default value if FALSE.}

\item{gene_num_gezero}{Input (Optional). A threshold (interger) to determine the inclusion of a gene.
The gene included needs to be expressed in at least \emph{gene_num_gezero} cells.
The default value is \emph{gene_num_gezero} = 3.}

\item{cell_num_gezero}{Input (Optional). A threshold (interger) to determine the inclusion of a cell.
The cell included needs to contain at least \emph{cell_num_gezero} expressed genes.
The default value is \emph{cell_num_gezero} = 10.}

\item{numGeneforEst}{Input (Optional). Use top \emph{numGeneforEst} (integer) highly expressed genes to estimate scaling factors. The default value is \emph{numGeneforEst} = 2000.}

\item{divideforFast}{Input (Optional). A logic value to indicate if speeding up computation
by randomly dividing cells into \emph{numDivide} groups. It is recommended to use for a large
number of cells, for example, > 50K cells. The default value is FALSE. Please input an integer in 
\emph{numDivide} below if \emph{divideforFast} = TRUE}

\item{numDivide}{Input (Optional). An integer is required if \emph{divideforFast} = TRUE. Suggest \emph{numDivide} = #cells/10K.}

}
\value{
\item{NormalizedData}{Matrix (genes by cells). Data matrix after normalization.}

\item{scalingFactor}{Vector. Cell-specific scaling factors.}

\item{delete_genes}{Vector. Indeice of the genes deleted.}

\item{delete_cells}{Vector. Indeice of the cells deleted.}
}
\description{
A function of normalizing single cell RNA-seq gene expression.
}
\examples{
set.seed(12345)
G <- 2000; n <- 600 # G: number of genes, n: number of cells
NB_cell <- function(j) rnbinom(G, size = 0.1, mu = rgamma(G, shape = 2, rate = 2))
countsimdata <- sapply(1:n, NB_cell)
colnames(countsimdata) <- paste("cell", 1:n, sep = "_")
rownames(countsimdata) <- paste("gene", 1:G, sep = "_")
Result <- LocASN(countmatrix = as(countsimdata,"sparseMatrix"))
Result$NormalizedData[1:10,1:10]; Result$scalingFactor[1:10]

#conditions <- c(rep(1,n/2), rep(2,n/2))
#Result2 <- LocASN(countmatrix = countsimdata, conditions = conditions)
#Result2$NormalizedData[1:10,1:10]; Result2$scalingFactor[1:10]
}
