% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocASN.R
\name{LocASN}
\alias{LocASN}
\title{Single-cell RNA sequencing normalization using a local average technique}
\usage{
LocASN(countmatrix, conditions = NULL, gene_num_gezero = 3,
  cell_num_gezero = 10, cpucores = 2L, numforEst = 10)
}
\arguments{
\item{countmatrix}{Input. Unnormalized count matrix (genes by cells).}

\item{conditions}{Input (Optional). Indicate which cells are sampled from the same conditions.
The default value, NULL, denotes all the cells are sampled from the same condition.}

\item{gene_num_gezero}{Input (Optional). A threshold (interger) to determine the inclusion of a gene.
The gene included needs to be expressed in at least \emph{gene_num_gezero} cells.
The default value is \emph{gene_num_gezero} = 3.}

\item{cell_num_gezero}{Input (Optional). A threshold (interger) to determine the inclusion of a cell.
The cell included needs to contain at least \emph{cell_num_gezero} expressed genes.
The default value is \emph{cell_num_gezero} = 10.}

\item{cpucores}{Input (Optional). Use \emph{cpucores} (integer) CPUs for parallel computation.
The default value is \emph{cpucores} = 2L.}

\item{numforEst}{Input (Optional). Use genes which expressed in at least \emph{numforEst} (integer) cells
to calculate the similarity between cells.
The default value is \emph{numforEst} = 10.}
}
\value{
\item{NormalizedData}{Matrix (genes by cells). Data matrix after normalization.}

\item{scalingFactor}{Vector. Cell-specific scaling factors.}

\item{delete_genes}{Vector. Indeice of the genes deleted.}

\item{delete_cells}{Vector. Indeice of the cells deleted.}
}
\description{
A function of normalizing single cell RNA-seq gene expression.
}
\examples{
set.seed(12345)
G <- 2000; n <- 600 # G: number of genes, n: number of cells
NB_cell <- function(j) rnbinom(G, size = 0.1, mu = rgamma(G, shape = 2, rate = 2))
countsimdata <- sapply(1:n, NB_cell)
colnames(countsimdata) <- paste("cell", 1:n, sep = "_")
rownames(countsimdata) <- paste("gene", 1:G, sep = "_")
Result <- LocASN(countmatrix = countsimdata)
Result$NormalizedData[1:10,1:10]; Result$scalingFactor[1:10]

#conditions <- c(rep(1,n/2), rep(2,n/2))
#Result2 <- LocASN(countmatrix = countsimdata, conditions = conditions, cpucores = 3L)
#Result2$NormalizedData[1:10,1:10]; Result2$scalingFactor[1:10]
}
